chore: use react compiler
