+++ frontend/package.json
@@ -42,6 +42,7 @@
     "@typescript-eslint/eslint-plugin": "^8.24.1",
     "@typescript-eslint/parser": "^8.24.1",
     "@vitejs/plugin-react": "^5.0.2",
+    "babel-plugin-react-compiler": "^1.0.0",
     "baseline-browser-mapping": "^2.9.7",
     "eslint": "^9.36.0",
     "eslint-config-prettier": "^10.0.1",
@@ -54,6 +55,9 @@
     "prettier": "^3.5.1",
     "typescript": "~5.9.2",
     "typescript-eslint": "^8.44.0",
-    "vite": "^7.1.4"
+    "vite": "npm:rolldown-vite@7.1.12"
+  },
+  "resolutions": {
+    "vite": "npm:rolldown-vite@7.1.12"
   }
 }
